Move validation dir to tests dir
