+++ xsc_test.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/jfrog/jfrog-cli-core/v2/common/format"
 
+	"github.com/jfrog/jfrog-cli-security/tests/validations"
 	"github.com/jfrog/jfrog-cli-security/utils/formats"
-	"github.com/jfrog/jfrog-cli-security/utils/validations"
 	"github.com/jfrog/jfrog-cli-security/utils/xsc"
 
 	"github.com/jfrog/jfrog-cli-security/tests"
